Feat/#22 로그인 로그아웃 기능 구현

+++ backend/src/main/java/com/team5/backend/domain/category/service/CategoryService.java
@@ -41,13 +41,13 @@ public class CategoryService {
                 .collect(Collectors.toList());
     }
 
-    public CategoryResDto getCategoryById(Integer categoryId) {
+    public CategoryResDto getCategoryById(Long categoryId) {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found with id " + categoryId));
         return toResponse(category);
     }
 
-    public CategoryResDto updateCategory(Integer categoryId, CategoryUpdateReqDto request) {
+    public CategoryResDto updateCategory(Long categoryId, CategoryUpdateReqDto request) {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found with id " + categoryId));
 
@@ -57,7 +57,7 @@ public class CategoryService {
         return toResponse(updated);
     }
 
-    public void deleteCategory(Integer categoryId) {
+    public void deleteCategory(Long categoryId) {
         categoryRepository.deleteById(categoryId);
     }
